I has done sooo much here

+++ src/net/minecraft/block/natural/BlockSoulSand.java
@@ -1,7 +1,7 @@
 package net.minecraft.block.natural;
 
 import net.minecraft.block.core.Block;
-import net.minecraft.src.AxisAlignedBB;
+import net.minecraft.misc.AxisAlignedBB;
 import net.minecraft.entity.Entity;
 import net.minecraft.block.material.Material;
 import net.minecraft.world.World;
@@ -13,7 +13,7 @@ public class BlockSoulSand extends Block {
 
 	public AxisAlignedBB getCollisionBoundingBoxFromPool(World var1, int var2, int var3, int var4) {
 		float var5 = 2.0F / 16.0F;
-		return AxisAlignedBB.getBoundingBoxFromPool((double)var2, (double)var3, (double)var4, (double)(var2 + 1), (double)((float)(var3 + 1) - var5), (double)(var4 + 1));
+		return AxisAlignedBB.getBoundingBoxFromPool(var2, var3, var4, var2 + 1, (float)(var3 + 1) - var5, var4 + 1);
 	}
 
 	public void onEntityCollidedWithBlock(World var1, int var2, int var3, int var4, Entity var5) {
